change some data type

+++ src/api/futures/public/instruments/mod.rs
@@ -46,11 +46,11 @@ pub struct MarginLevel {
 pub struct Instrument {
     pub symbol: String,
     pub category: Option<String>,
-    pub contract_size: Option<f64>,
-    pub contract_value_trade_precision: Option<f64>,
+    pub contract_size: Option<u32>,
+    pub contract_value_trade_precision: Option<u32>,
     pub fee_schedule_uid: Option<String>,
-    pub funding_rate_coefficient: Option<f64>,
-    pub impact_mid_size: Option<f64>,
+    pub funding_rate_coefficient: Option<u32>,
+    pub impact_mid_size: Option<u32>,
     pub isin: Option<String>,
     pub last_trading_time: Option<String>,
     #[serde(default)]
